requirements.txt
flask
paho-mqtt
pandas
numpy
plotly